% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{collect_indices}
\alias{collect_indices}
\title{Collect polynomial constraints that characterize the semialgebraic variety of a given tree}
\usage{
collect_indices(g, m, nr_4 = NULL, nr_3 = NULL)
}
\arguments{
\item{g}{An igraph object that is a tree. It is assumed that the first m nodes correspond to oberseved nodes.}

\item{m}{Integer, number of observed nodes.}

\item{nr_4}{Number of considered subsets of size 4. This is optional. 
If \code{NULL}, all subsets are considered. 
If a number is given, subsets are choosen randomly by the function \code{\link{random_combs}}.}

\item{nr_3}{Number of considered subsets of size 3. This is optional. 
If \code{NULL}, all subsets are considered. 
If a number is given, subsets are choosen randomly by the function \code{\link{random_combs}}.}
}
\value{
A list with three entries is returned. All entries are matrices. 
The first matrx contains the equality constraints and has 8 columns. 
The second matrx contains all inequality constraints where only 6 indices are necessary (i.e. it has 6 columns). 
The third matrx contains the inequality constraints where 8 indices are necessary (i.e. it has 8 columns).
}
\description{
The polynomials are represented as a list of indices. Example:
If one polynomial is \eqn{\sigma_{pq} \sigma_{rs} - \sigma_{pr} \sigma_{qs}} then the indices will be \code{p,q,r,s,p,r,q,s}.
}
\examples{
vertices <- data.frame(name=seq(1,8), type=c(rep(1,5), rep(2,3))) # 1=observed, 2=latent
edges <- data.frame(from=c(1,2,3,4,5,6,7), to=c(8,8,6,6,7,7,8))
tree <- igraph::graph_from_data_frame(edges, directed=FALSE, vertices=vertices)
plot(tree)
collect_indices(tree, 5)
}
