% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-functions.R
\name{findQ}
\alias{findQ}
\title{Determine the set Q}
\usage{
findQ(g, m, nr = NULL)
}
\arguments{
\item{g}{An igraph object that is a tree. It is assumed that the first m nodes correspond to oberseved nodes.}

\item{m}{Integer, number of observed nodes.}

\item{nr}{Integer, number of subsets that are considered. Note that this is optional. If \code{NULL}, all subsets are considered.}
}
\value{
A list with two entries is returned. Both entries are matrices with 4 columns. 
The first matrix are all subsets that belong to Q, the second are all subsets that do not belong to Q. 
Each row represents one subset. For the matrix Q the order of the elements in a row is important. 
A row with entries p,q,r,s means that the element {p,q}|{r,s} is in Q.
}
\description{
For all subsets of size 4 of the observed nodes of a tree, the function determines if a subset is an elemnt of Q or not.
}
\examples{
vertices <- data.frame(name=seq(1,8), type=c(rep(1,5), rep(2,3))) # 1=observed, 2=latent
edges <- data.frame(from=c(1,2,3,4,5,6,7), to=c(8,8,6,6,7,7,8))
tree <- igraph::graph_from_data_frame(edges, directed=FALSE, vertices=vertices)
plot(tree)
findQ(tree, 5)
}
