% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests-factoranalysis.R
\name{get_minors}
\alias{get_minors}
\title{Get all possible minors in a factor analysis model}
\usage{
get_minors(m, factors)
}
\arguments{
\item{m}{Integer, number of observed nodes}

\item{factors}{Integer, number of hidden factors.}
}
\value{
Matrix with \code{2*factors+2} columns. Each row corresponds to one (factors+1)x(factors+1)-minor.
The first \code{factors+1} columns correspond to the row-indices of the minor, 
the last \code{factors+1} columns correspond to the column-indices of the minor.
}
\description{
Get all possible minors in a factor analysis model
}
\examples{
get_minors(8,2)
}
