% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests-factoranalysis.R
\name{test_indep_factors}
\alias{test_indep_factors}
\title{Tests the (3x3)-minor constraints of a 2-factor model by the maximum of a high-dimensional independent sum}
\usage{
test_indep_factors(X, nr_minors, E = 1000)
}
\arguments{
\item{X}{Matrix with observed data. The number of columns corresponds to the number of observed variables. 
Each row corresponds to one sample.}

\item{nr_minors}{Integer, number of randomly chosen minors that are tested.}

\item{E}{Integer, number of bootstrap iterations.}
}
\value{
Named list with two entries: Test statistic (\code{TSTAT}) and p-value (\code{PVAL}).
}
\description{
This function tests the (3x3)-minor constraints of a given 2-factor model to observed data.
The minors are estimated by grouping the data into independent subsets. 
Each group is used to form an unbiased estimate of all minors. 
The test statistic is the maximum of the average of the independent studentized estimates. 
A Gaussian multiplier bootstrap procedure is used to estimate the limiting distribution of 
the test statistic and to compute the p-value of the test.
}
\examples{
# Covariance matrix corresponding to the two-factor analysis model
m=20
Gamma = matrix(stats::rnorm(2*m),m,2)
Psi = diag(rep(1,m))
cov = Psi + Gamma \%*\% t(Gamma)

# Sample data from the two-factor analysis model
X = MASS::mvrnorm(500, mu=rep(0,nrow(cov)), Sigma=cov)

# Apply the test
test_indep_factors(X, 10000)

}
\references{
TO BE WRITTEN
}
